third_party/blink/web_tests/external/wpt/css/cssom/cssstyledeclaration-all-shorthand-expected.txt

This is a testharness.js-based test.
[FAIL] getPropertyValue('all') returns empty string when single property overriden
  assert_equals: expected "" but got "revert"
[FAIL] setProperty('all') sets all property values
  assert_equals: expected "revert" but got ""
Harness: the test ran to completion.
